pyproject.toml
[project]
name = "open-ticket-ai"
version = "0.1.0"
description = "AI enhancements for open source ticket systems"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub~=0.33.0",
    "pydantic~=2.11.7",
    "pydantic_core",
    "typing_extensions",
    "spacy",
    "Faker",
    "phonenumbers",
    "pytest~=8.4.0",
    "pluggy",
    "pygments",
    "iniconfig",
    "pdoc3",
    "PyYAML~=6.0.2",
    "transformers~=4.52.4",
    "injector~=0.22.0",
    "schedule~=1.2.2",
    "requests~=2.32.4",
    "pyfiglet>=1.0.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.29.1",
    "mkdocs-gen-files>=0.5.0",
    "docstring-parser>=0.16",
    "openai>=1.93.0",
    "tenacity>=9.1.2",
    "prefect>=3.4.8",
    "otobo>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
include = ["open_ticket_ai*"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
# Add specific lint configurations here if needed in the future

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
